+++ amplipi/streams.py
@@ -21,36 +21,39 @@ such as Pandora, Spotify, and AirPlay. Each digital source is expected to have
 a consistent interface.
 """
 
-# Used by InternetRadio and Spotify
-import json
 import os
+import sys
+import subprocess
+import time
+from typing import Union, Optional, List
+import threading
+
+import ast
+import json
 import signal
 import socket
 import hashlib # md5 for string -> MAC generation
-import subprocess
-import sys
-import threading
-import time
-from typing import Union, Optional
 
-import amplipi.models as models
-import amplipi.utils as utils
+from amplipi import models
+from amplipi import utils
 from amplipi.mpris import MPRIS
 
+# We use Popen for long running process control this error is not useful:
+# pylint: disable=consider-using-with
 
 def write_config_file(filename, config):
   """ Write a simple config file (@filename) with key=value pairs given by @config """
-  with open(filename, 'wt') as cfg_file:
+  with open(filename, 'wt', encoding='utf-8') as cfg_file:
     for key, value in config.items():
       cfg_file.write(f'{key}={value}\n')
 
 def write_sp_config_file(filename, config):
   """ Write a shairport config file (@filename) with a hierarchy of grouped key=value pairs given by @config """
-  with open(filename, 'wt') as cfg_file:
+  with open(filename, 'wt', encoding='utf-8') as cfg_file:
     for group, gconfig in config.items():
       cfg_file.write(f'{group} =\n{{\n')
       for key, value in gconfig.items():
-        if type(value) is str:
+        if isinstance(value, str):
           cfg_file.write(f'  {key} = "{value}"\n')
         else:
           cfg_file.write(f'  {key} = {value}\n')
@@ -58,8 +61,7 @@ def write_sp_config_file(filename, config):
 
 def uuid_gen():
   """ Generates a UUID for use in DLNA and Plexamp streams """
-  u_args = 'uuidgen'
-  uuid_proc = subprocess.run(args=u_args, capture_output=True)
+  uuid_proc = subprocess.run(args='uuidgen', capture_output=True, check=False)
   uuid_str = str(uuid_proc).split(',')
   c_check = uuid_str[0]
   val = uuid_str[2]
@@ -74,12 +76,15 @@ class BaseStream:
   def __init__(self, stype: str, name: str, only_src=None, disabled: bool=False, mock=False):
     self.name = name
     self.disabled = disabled
-    self.proc = None
+    self.proc: Optional[subprocess.Popen] = None
     self.mock = mock
-    self.src = None
-    self.only_src = only_src
+    self.src: Optional[int] = None
+    self.only_src: Optional[int] = only_src
     self.state = 'disconnected'
     self.stype = stype
+
+  def __del__(self):
+    self.disconnect()
 
   def __str__(self):
     connection = f' connected to src={self.src}' if self.src else ''
@@ -95,7 +100,7 @@ class BaseStream:
     return f'{self.name} - {self.stype}'
 
   def _disconnect(self):
-    print(f'{self.name} disconnected', flush=True)
+    print(f'{self.name} disconnected')
     self.state = 'disconnected'
     self.src = None
 
@@ -109,9 +114,12 @@ class BaseStream:
     self._disconnect()
 
   def _connect(self, src):
-    print(f'{self.name} connected to {src}', flush=True)
+    print(f'{self.name} connected to {src}')
     self.state = 'connected'
     self.src = src
+
+  def is_connected(self) -> bool:
+    return self.src is not None
 
   def connect(self, src: int):
     """ Connect the stream to an output source """
@@ -373,9 +381,9 @@ class Spotify(BaseStream):
   def __init__(self, name: str, disabled: bool = False, mock: bool = False):
     super().__init__('spotify', name, disabled=disabled, mock=mock)
 
-    self.connect_port = None
-    self.mpris = None
-    self.proc_pid = None
+    self.connect_port: Optional[int] = None
+    self.mpris: Optional[MPRIS] = None
+    self.proc_pid: Optional[int] = None
     self.supported_cmds = ['play', 'pause', 'next', 'prev']
 
   def reconfig(self, **kwargs):
@@ -525,8 +533,6 @@ class Pandora(BaseStream):
       time.sleep(0.1) # delay a bit, is this needed?
       self.connect(last_src)
 
-  def __del__(self):
-    self.disconnect()
 
   def connect(self, src):
     """ Connect pandora output to a given audio source
@@ -594,7 +600,7 @@ class Pandora(BaseStream):
       img_url='static/imgs/pandora.png'
     )
     try:
-      with open(loc, 'r') as file:
+      with open(loc, 'r', encoding='utf-8') as file:
         for line in file.readlines():
           line = line.strip()
           if line:
@@ -620,7 +626,7 @@ class Pandora(BaseStream):
     """
     try:
       if cmd in self.supported_cmds:
-        with open(self.ctrl, 'w') as file:
+        with open(self.ctrl, 'w', encoding='utf-8') as file:
           file.write(self.supported_cmds[cmd]['cmd'])
           file.flush()
         expected_state = self.supported_cmds[cmd]['state']
@@ -629,7 +635,7 @@ class Pandora(BaseStream):
       elif 'station' in cmd:
         station_id = int(cmd.replace('station=', ''))
         if station_id is not None:
-          with open(self.ctrl, 'w') as file:
+          with open(self.ctrl, 'w', encoding='utf-8') as file:
             file.write('s')
             file.flush()
             file.write(f'{station_id}\n')
@@ -704,11 +710,11 @@ class DLNA(BaseStream):
     loc = f'{src_config_folder}/currentSong'
     source = models.SourceInfo(name=self.full_name(), state=self.state, img_url='static/imgs/dlna.png')
     try:
-      with open(loc, 'r') as file:
+      with open(loc, 'r', encoding='utf-8') as file:
         for line in file.readlines():
           line = line.strip()
           if line:
-            d = eval(line)
+            d = ast.literal_eval(line)
         source.state = d['state']
         source.album = d['album']
         source.artist = d['artist']
@@ -740,9 +746,6 @@ class InternetRadio(BaseStream):
       self.disconnect()
       time.sleep(0.1) # delay a bit, is this needed?
       self.connect(last_src)
-
-  def __del__(self):
-    self.disconnect()
 
   def connect(self, src):
     """ Connect a VLC output to a given audio source
@@ -782,10 +785,12 @@ class InternetRadio(BaseStream):
     loc = f'{src_config_folder}/currentSong'
     source = models.SourceInfo(name=self.full_name(),
                               state=self.state,
-                              img_url=self.logo,
+                              img_url='static/imgs/internet_radio.png',
                               supported_cmds=self.supported_cmds)
+    if self.logo:
+      source.img_url = self.logo
     try:
-      with open(loc, 'r') as file:
+      with open(loc, 'r', encoding='utf-8') as file:
         data = json.loads(file.read())
         source.artist = data['artist']
         source.track = data['track']
@@ -818,26 +823,14 @@ class InternetRadio(BaseStream):
       pass
 
 class Plexamp(BaseStream):
-  """ A Plexamp Stream """
+  """ A Plexamp Stream
+  TODO: old plexamp interface was disabled, integrate support for new PlexAmp
+  """
   def __init__(self, name: str, client_id, token, disabled: bool = False, mock: bool = False):
     super().__init__('plexamp', name, disabled=disabled, mock=mock)
-    self.client_id = client_id
-    self.token = token
 
   def reconfig(self, **kwargs):
-    reconnect_needed = False
-    if 'name' in kwargs and kwargs['name'] != self.name:
-      self.name = kwargs['name']
-      reconnect_needed = True
-    if reconnect_needed:
-      if self._is_running():
-        last_src = self.src
-        self.disconnect()
-        time.sleep(0.1) # delay a bit, is this needed?
-        self.connect(last_src)
-
-  def __del__(self):
-    self.disconnect()
+    self.name = kwargs['name']
 
   def connect(self, src):
     """ Connect plexamp output to a given audio source
@@ -847,70 +840,12 @@ class Plexamp(BaseStream):
       self._connect(src)
       return
 
-    src_config_folder = f'{utils.get_folder("config")}/srcs/{src}'
-    mpd_template = f'{utils.get_folder("streams")}/mpd.conf'
-    plexamp_template = f'{utils.get_folder("streams")}/server.json'
-    plexamp_home = src_config_folder
-    plexamp_config_folder = f'{plexamp_home}/.config/Plexamp'
-    mpd_conf_file = f'{plexamp_home}/mpd.conf'
-    server_json_file = f'{plexamp_config_folder}/server.json'
-    # make all of the necessary dir(s)
-    os.system(f'mkdir -p {plexamp_config_folder}')
-    os.system(f'cp {mpd_template} {mpd_conf_file}')
-    os.system(f'cp {plexamp_template} {server_json_file}')
-    self.uuid = uuid_gen()
-    # server.json config (Proper server.json file must be copied over for this to work)
-    with open(server_json_file) as json_file:
-      contents = json.load(json_file)
-      r_token = contents['user']['token']
-      if r_token != '_': # Dummy value from template
-        self.token = r_token
-
-    # Double quotes required for Python -> JSON translation
-    json_config = {
-      "player": {
-        "name": f"{self.name}",
-        "identifier": f"{self.client_id}"
-      },
-      "user": {
-        "token": self.token
-      },
-      "state": "null",
-      "server": "null",
-      "audio": {
-        "normalize": "false",
-        "crossfade": "false",
-        "mpd_path": f"{mpd_conf_file}"
-      }
-    }
-
-    with open(server_json_file, 'w') as new_json:
-      json.dump(json_config, new_json, indent=2)
-
-    # mpd.conf creation
-    with open(mpd_conf_file, 'r') as MPD:
-      data = MPD.read()
-      data = data.replace('ch', f'ch{src}')
-      data = data.replace('GENERIC_LOGFILE_LOCATION', f'{plexamp_config_folder}/mpd.log')
-    with open(mpd_conf_file, 'w') as MPD:
-      MPD.write(data)
-    # PROCESS
-    plexamp_args = ['/usr/bin/node', '/home/pi/plexamp/server/server.prod.js']
-    try:
-      self.proc = subprocess.Popen(args=plexamp_args, preexec_fn=os.setpgrp, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, env={'HOME' : plexamp_home})
-      time.sleep(0.1) # Delay a bit
-      self._connect(src)
-    except Exception as exc:
-      print(f'error starting plexamp: {exc}')
-
   def disconnect(self):
-    if self._is_running():
-      os.killpg(os.getpgid(self.proc.pid), signal.SIGKILL)
     self._disconnect()
-    self.proc = None
 
   def info(self) -> models.SourceInfo:
     source = models.SourceInfo(name=self.full_name(), state=self.state, img_url='static/imgs/plexamp.png')
+    source.track = "Not currently supported"
     return source
 
 class FilePlayer(BaseStream):
@@ -919,9 +854,6 @@ class FilePlayer(BaseStream):
     super().__init__('file player', name, disabled=disabled, mock=mock)
     self.url = url
     self.bkg_thread = None
-
-  def __del__(self):
-    self.disconnect()
 
   def reconfig(self, **kwargs):
     reconnect_needed = False
@@ -988,7 +920,7 @@ class FMRadio(BaseStream):
     reconnect_needed = False
     ir_fields = ['freq', 'logo']
     fields = list(ir_fields) + ['name']
-    for k,v in kwargs.items():
+    for k, v in kwargs.items():
       if k in fields and self.__dict__[k] != v:
         self.__dict__[k] = v
         if k in ir_fields:
@@ -998,9 +930,6 @@ class FMRadio(BaseStream):
       self.disconnect()
       time.sleep(0.1) # delay a bit, is this needed?
       self.connect(last_src)
-
-  def __del__(self):
-    self.disconnect()
 
   def connect(self, src):
     """ Connect a fmradio.py output to a given audio source """
@@ -1038,7 +967,7 @@ class FMRadio(BaseStream):
       self.logo = "static/imgs/fmradio.png"
     source = models.SourceInfo(name=self.full_name(), state=self.state, img_url=self.logo)
     try:
-      with open(loc, 'r') as file:
+      with open(loc, 'r', encoding='utf-8') as file:
         data = json.loads(file.read())
         # Example JSON: "station": "Mixx96.1", "callsign": "KXXO", "prog_type": "Soft rock", "radiotext": "        x96.1"
         #print(json.dumps(data))
@@ -1085,9 +1014,6 @@ class LMS(BaseStream):
         self.disconnect()
         time.sleep(0.1) # delay a bit, is this needed?
         self.connect(last_src)
-
-  def __del__(self):
-    self.disconnect()
 
   def connect(self, src):
     """ Connect a sqeezelite output to a given audio source
@@ -1162,22 +1088,22 @@ def build_stream(stream: models.Stream, mock=False) -> AnyStream:
   disabled = args.pop('disabled', False)
   if stream.type == 'rca':
     return RCA(name, args['index'], disabled=disabled, mock=mock)
-  elif stream.type == 'pandora':
+  if stream.type == 'pandora':
     return Pandora(name, args['user'], args['password'], station=args.get('station', None), disabled=disabled, mock=mock)
-  elif stream.type in ['shairport', 'airplay']: # handle older configs
+  if stream.type in ['shairport', 'airplay']: # handle older configs
     return AirPlay(name, args.get('ap2', False), disabled=disabled, mock=mock)
-  elif stream.type == 'spotify':
+  if stream.type == 'spotify':
     return Spotify(name, disabled=disabled, mock=mock)
-  elif stream.type == 'dlna':
+  if stream.type == 'dlna':
     return DLNA(name, disabled=disabled, mock=mock)
-  elif stream.type == 'internetradio':
+  if stream.type == 'internetradio':
     return InternetRadio(name, args['url'], args.get('logo'), disabled=disabled, mock=mock)
-  elif stream.type == 'plexamp':
+  if stream.type == 'plexamp':
     return Plexamp(name, args['client_id'], args['token'], disabled=disabled, mock=mock)
-  elif stream.type == 'fileplayer':
+  if stream.type == 'fileplayer':
     return FilePlayer(name, args['url'], disabled=disabled, mock=mock)
-  elif stream.type == 'fmradio':
+  if stream.type == 'fmradio':
     return FMRadio(name, args['freq'], args.get('logo'), disabled=disabled, mock=mock)
-  elif stream.type == 'lms':
+  if stream.type == 'lms':
     return LMS(name, args.get('server'), disabled=disabled, mock=mock)
   raise NotImplementedError(stream.type)